duplicate user protection

+++ profile/create.php
@@ -16,14 +16,20 @@ $data = json_decode(file_get_contents("php://input"));
 $db = new Database(); 
 $user = new User($db->pdo, $data->firstname, $data->lastname, $data->email, $data->seminar, $data->password); 
 
-if($user->createDBEntry()) {
-    // case: user created
-    http_response_code(200);
-    echo json_encode(array("message" => "User was created."));
+// check if the user already exists
+if ($db->emailExists($user->getEmail()) == null) {
+    if($user->createDBEntry()) {
+        // case: user created
+        http_response_code(200);
+        echo json_encode(array("message" => "User was created."));
+    } else {
+        // case: user creation failed
+        http_response_code(400);
+        echo json_encode(array("message" => "Unable to create user."));
+    }
 } else {
-    // case: user creation failed
-    http_response_code(400);
-    echo json_encode(array("message" => "Unable to create user."));
+    http_response_code(200);
+    echo json_encode(array("message" => "User already exists."));
 }
 ?>
